add and fixes utils

+++ dlstats/utils.py
@@ -10,18 +10,32 @@ import traceback
 
 import requests
 
-try:
-    import requests_cache
-    HAVE_REQUESTS_CACHE = True
-except ImportError:
-    HAVE_REQUESTS_CACHE = False
-
 logger = logging.getLogger(__name__)
 
 def last_error():
     f = StringIO() 
     traceback.print_exc(file=f)
     return f.getvalue()
+
+def make_store_path(base_path=None, provider_name=None, dataset_code=None):
+    store_filepath = None
+    
+    if not base_path:
+        store_filepath = tempfile.mkdtemp()
+    else:
+        store_filepath = base_path
+    
+    if provider_name:
+        store_filepath = os.path.abspath(os.path.join(store_filepath, provider_name))
+        if not os.path.exists(store_filepath):
+            os.makedirs(store_filepath)
+    
+    if dataset_code:
+        store_filepath = os.path.abspath(os.path.join(store_filepath, dataset_code))
+        if not os.path.exists(store_filepath):
+            os.makedirs(store_filepath)
+    
+    return store_filepath
 
 class Downloader:
     
@@ -32,8 +46,7 @@ class Downloader:
     def __init__(self, url=None, filename=None, store_filepath=None, 
                  timeout=None, max_retries=0, 
                  replace=True, force_replace=True,
-                 headers={},
-                 cache=None):
+                 headers={}, client=None):
         
         self.url = url
         self.filename = filename
@@ -42,6 +55,13 @@ class Downloader:
         self.max_retries = max_retries
         self.force_replace = force_replace
         self.headers = headers
+        self.client = client or requests
+
+        if not self.url:
+            raise ValueError("url is required")
+        
+        if not self.filename:
+            raise ValueError("filename is required")
         
         for item in self.DEFAULT_HEADERS.items():
             self.headers.setdefault(*item)
@@ -57,9 +77,6 @@ class Downloader:
         if os.path.exists(self.filepath) and not replace:
             raise Exception("filepath is already exist : %s" % self.filepath)
 
-        if HAVE_REQUESTS_CACHE and cache:
-            requests_cache.install_cache(**cache)
-        
     def _download(self, raise_errors=True):
         
         #TODO: max_retries (self.max_retries)
@@ -67,7 +84,7 @@ class Downloader:
         
         start = time.time()
         try:
-            response = requests.get(self.url, 
+            response = self.client.get(self.url, 
                                     timeout=self.timeout, 
                                     stream=True,
                                     allow_redirects=True,
@@ -76,7 +93,7 @@ class Downloader:
 
             code = int(response.status_code)
             
-            if code >= 400:
+            if code == 304 or code >= 400:
                 msg = "download url[%s] - status_code[%s] - reason[%s]" % (self.url, 
                                                                            code, 
                                                                            response.reason)
@@ -87,24 +104,15 @@ class Downloader:
                     logger.warning(msg)
                     return response
 
-            with open(self.filepath, 'wb') as f:
+            with open(self.filepath, mode='wb') as f:
                 for chunk in response.iter_content():
                     f.write(chunk)
 
             return response
         
-            """
-            except requests.exceptions.ConnectionError as err:
-                raise Exception("Connection Error")
-            except requests.exceptions.ConnectTimeout as err:
-                raise Exception("Connect Timeout")
-            except requests.exceptions.ReadTimeout as err:
-                raise Exception("Read Timeout")
-            """
         except Exception as err:
             logger.critical("Not captured exception : %s" % str(err))
             raise
-            #raise Exception("Not captured exception : %s" % str(err))            
 
         end = time.time() - start
         logger.info("download file[%s] - END - time[%.3f seconds]" % (self.url, end))
@@ -115,10 +123,10 @@ class Downloader:
             os.remove(self.filepath)
         
         if not os.path.exists(self.filepath):
-            logger.info("not found file[%s] - download dataset url[%s]" % (self.filepath, self.url))
+            logger.warning("not found file[%s] - download dataset url[%s]" % (self.filepath, self.url))
             self._download()
         else:
-            logger.info("use local dataset file [%s]" % self.filepath)
+            logger.warning("use local dataset file [%s]" % self.filepath)
         
         return self.filepath
 
@@ -130,10 +138,10 @@ class Downloader:
             os.remove(self.filepath)
         
         if not os.path.exists(self.filepath):
-            logger.info("not found file[%s] - download dataset url[%s]" % (self.filepath, self.url))
+            logger.warning("not found file[%s] - download dataset url[%s]" % (self.filepath, self.url))
             response = self._download(raise_errors=False)
         else:
-            logger.info("use local dataset file [%s]" % self.filepath)
+            logger.warning("use local dataset file [%s]" % self.filepath)
         
         return self.filepath, response
 
@@ -167,10 +175,11 @@ def clean_datetime(dt=None,
         tzinfo = None
     return datetime(year, month, day, hour, minute, second, microsecond, tzinfo=tzinfo)
     
-def remove_file_and_dir(filepath):
+def remove_file_and_dir(filepath, let_root=False):
     if not os.path.exists(filepath):
         logger.warning("file not found [%s]" % filepath)
         return
+    
     if not os.path.isfile(filepath):
         logger.warning("file is not file [%s]" % filepath)
         return
@@ -187,6 +196,50 @@ def remove_file_and_dir(filepath):
     if logger.isEnabledFor(logging.DEBUG):
         logger.debug("remove file[%s]" % filepath)    
     os.remove(filepath)
+    
     if logger.isEnabledFor(logging.DEBUG):
-        logger.debug("remove dir[%s]" % dirname)    
-    os.rmdir(dirname)
+        logger.debug("remove dir[%s]" % dirname)
+
+    for root, dirs, files in os.walk(dirname, topdown=False):
+        for name in files:
+            _filepath = os.path.join(root, name)
+            try:
+                os.remove(_filepath)
+                logger.debug("remove file[%s]" % _filepath)
+            except Exception as err:
+                logger.error("not remove file[%s] - error[%s]" % (_filepath, str(err)))
+        for name in dirs:
+            _filepath = os.path.join(root, name)
+            try:
+                os.rmdir(_filepath)            
+                logger.debug("remove dir[%s]" % _filepath)
+            except Exception as err:
+                logger.error("not remove dir[%s] - error[%s]" % (_filepath, str(err)))
+
+    if os.path.exists(dirname):
+        if let_root:
+            logger.debug("root dir not remove %s" % dirname)
+        else:
+            try:
+                os.rmdir(dirname)
+            except Exception as err:
+                logger.error("root dir not remove %s" % dirname)
+
+
+def get_ordinal_from_period(date_str, freq=None):
+    from dlstats.cache import cache
+    from dlstats import constants
+    import pandas
+    
+    if not cache or not freq in constants.CACHE_FREQUENCY:
+        return pandas.Period(date_str, freq=freq).ordinal
+    
+    key = "%s.%s" % (date_str, freq)
+    period_from_cache = cache.get(key)
+    if period_from_cache:
+        return period_from_cache
+    
+    period_ordinal = pandas.Period(date_str, freq=freq).ordinal
+    cache.set(key, period_ordinal)
+    
+    return period_ordinal
